or1k: fix pthread_arch.h
The initial 'implementation' was just a copy of the
microblaze one, with the tp-reg swapped out to r10.

This changes this to reflect how TLS actually is implemented
on or1k (Variant I), with the TLS segment located after tp.
Our TCB only contains a pointer to DTV,
so the size of TCB = 4, and tp is located after TCB.

The way the tp reg (r10) is obtained is also slightly optimized
for the non-clang case.